Refactor retrieval of route name
This change refactors the code for retrieving the route's name into a
standalone function. My intent here is to make it easier to maintain and
test, as I have a number of ideas for how it can be improved in future
iterations, and this refactoring makes it easier to do than an inline
code block.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>