Update library to use latest versions of underlying packages and PHP 7.4
This package hasn't been updated in a while, so this is a more than
modest size change. It updates all of the package dependencies to use
the most recent versions, also accommodating name changes (Zend
Expressive -> Laminas). It also updates a lot of the code to reflect
more recent changes in PHP 7.4.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>